Fix typing errors and add back sample.env
